Add users, login, auth tokens, middleware extractors, user tests

+++ part4/bloglist/controllers/blogs.js
@@ -2,17 +2,39 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+  const user = request.user
+
+  const blog = new Blog({
+    user: user.id,
+    ...request.body
+  })
+
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
+blogsRouter.get('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  response.json(blog)
+})
+
 blogsRouter.delete('/:id', async (request, response) => {
+  const user = request.user
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== user.id) {
+    return response.status(401).json({ error : 'id mismatch' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
